floatchai ai/floatchat/backend/requirements.txt
# FloatChat Backend - Feature 1: Data Ingestion Pipeline
# Python 3.11+

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database & ORM
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
alembic==1.13.1
geoalchemy2==0.15.1

# Task Queue
celery==5.4.0
redis==5.0.4

# NetCDF & Data Processing
xarray==2024.3.0
netCDF4==1.6.5
numpy==1.26.4
pandas==2.2.2

# Object Storage
boto3==1.34.84

# Logging & Monitoring
structlog==24.1.0
sentry-sdk[fastapi]==1.45.0
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
python-logging-loki==0.3.1

# Configuration
pydantic-settings==2.2.1

# LLM Integration
openai==1.23.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
slowapi==0.1.9

# Geometry (for PostGIS)
shapely==2.1.2

# Vector Store (Feature 3: Metadata Search Engine)
pgvector>=0.2.5

# SQL Parsing & Validation (Feature 4: NL Query Engine)
sqlglot>=20.0.0

# Testing
pytest==8.1.2
pytest-asyncio==0.23.6
httpx==0.27.0
